+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added CHANGELOG.md
 - Added support for asserting a Task's Credentials/RoleArn field
 - Added support for asserting a Map's MaxConcurrency field
+- Added support for Map's ItemProcessor and ItemSelector fields
 
 ## 1.0.0